requirements.txt
numpy
rich
Pillow
click
tensorboard
absl-py
numpy
torch
torchvision # for quick video preprocessing
pytorch-lightning
av # torchvision requirement
megatron-lm